# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ulo \u2014 House of Stories \u2014 Mobile API",
    "version": "1.0.0",
    "description": "Sanctum bearer JSON for Expo mobile (AGENTS.md.mobile + API.md). Base URL {APP_URL}/api/v1. Mirror of routes/api.php. Auth: Bearer <sanctum_token>. Public: GET /pricing, GET /share/rooms/{slug}.",
    "contact": {
      "name": "Ulo API"
    }
  },
  "servers": [
    {
      "url": "https://uloak.test/api/v1",
      "description": "Local Herd (see get-absolute-url)"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/billing/checkout": {
      "post": {
        "tags": [
          "Billing"
        ],
        "summary": "Create checkout (server-side price, idempotent)",
        "operationId": "createCheckout",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Checkout",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Payment"
                    },
                    "authorization_url": {
                      "type": "string",
                      "format": "uri"
                    },
                    "reference": {
                      "type": "string"
                    },
                    "free": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "402": {
            "description": "Requires checkout"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Sanctum"
      }
    },
    "schemas": {
...
      "Payment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "successful",
              "failed"
            ]
          },
          "provider": {
            "type": "string",
            "enum": [
              "paystack",
              "stripe",
              "paypal"
            ]
          },
          "tier": {
            "type": [
              "string",
              "null"
            ],
            "description": "purchased tier key (room tier or subscription tier)"
          },
          "creator_profile_id": {
            "type": [
              "integer",
              "null"
            ],
            "description": "attributed creator when checkout used a creator ref_code"
          },
          "subscription_id": {
            "type": [
              "integer",
              "null"
            ],
            "description": "created subscription for subscription-tier payments"
          },
          "amount": {
            "type": "integer",
            "description": "minor units (kobo/cents)"
          },
          "currency": {
            "type": "string",
            "enum": [
              "NGN",
              "USD",
              "GBP",
              "EUR"
            ]
          },
          "region": {
            "type": [
              "string",
              "null"
            ]
          },
          "provider_reference": {
            "type": [
              "string",
              "null"
            ]
          },
          "room": {
            "type": [
              "object",
              "null"
            ],
            "properties": {
              "id": {
                "type": "integer"
              },
              "slug": {
                "type": "string"
              },
              "name": {
                "type": "string"
              }
            }
          },
          "paid_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "CheckoutRequest": {
        "type": "object",
        "required": [
          "region",
          "tier"
        ],
        "properties": {
          "room_id": {
            "type": "integer"
          },
          "region": {
            "type": "string",
            "enum": [
              "nigeria",
              "rest_of_africa",
              "uk",
              "us_rest_of_world",
              "europe"
            ]
          },
          "tier": {
            "type": "string",
            "enum": [
              "starter",
              "full_room",
              "family_archive",
              "family_monthly",
              "family_yearly",
              "viewer_monthly",
              "viewer_yearly",
              "viewer_vip_monthly",
              "viewer_vip_yearly"
            ]
          },
          "provider": {
            "type": "string",
            "enum": [
              "paystack",
              "paypal",
              "stripe"
            ]
          },
          "ref_code": {
            "type": "string",
            "description": "partner code or creator ref_code (creator attribution accrues a creator_earnings row on activation)"
          }
        }
      },
...
    }
  }
}